+++ src/models/index.ts
@@ -13,6 +13,7 @@ export * from './DataType';
 export * from './DraftField';
 export * from './Framework';
 export * from './GitSettings';
+export * from './LoadingType';
 export * from './MediaPaths';
 export * from './Mode';
 export * from './PanelSettings';